re work player mappings
