+++ src/lib/constants.ts
@@ -1,7 +1,7 @@
 import { $Enums } from "@prisma/client";
 import { getAllGames } from "../../prisma/lib/games";
 import { getAllMembers } from "../../prisma/lib/members";
-import { capitalizeFirst } from "./utils";
+import { PLAYER_MAPPINGS } from "../app/(routes)/admin/_utils/player-mappings";
 
 /**
  * Returns navigation data for all RDC members.
@@ -22,8 +22,8 @@ export const getMembersNav = async () => {
     desc: string;
     stats: { prop: string; val: string }[];
   }[] = members.data.map((member) => {
-    const memberKey = member.playerName.toLowerCase();
-    const rdcMember = RDCMembers.get(memberKey as MembersEnum)!;
+    const memberKey = member.playerName as keyof typeof PLAYER_MAPPINGS;
+    const rdcMember = PLAYER_MAPPINGS[memberKey]!;
     return {
       alt: rdcMember.nav.alt,
       name: member.playerName,
@@ -50,135 +50,6 @@ export enum MembersEnum {
   Ben = "ben",
   Des = "des",
 }
-
-const RDCMembers = new Map<MembersEnum, MembersProps>([
-  [
-    MembersEnum.Mark,
-    {
-      nav: {
-        alt: "RDC Mark",
-        name: "Cash Money Mawk",
-        url: "/members/mark",
-        src: "https://static.wikia.nocookie.net/rdcworld1/images/f/f2/Mark-Phillips.jpg/revision/latest/thumbnail/width/360/height/450?cb=20191004005953",
-      },
-      desc: "Director, animator, Mark is the founder and creator of RDC. He is one of the better gamers in RDC and plays as such.",
-      stats: [
-        { prop: "Best Rocket League Player", val: "#1" },
-        { prop: "Dragon Ball Sparking Zero", val: "#2" },
-        { prop: "Scariest Gamer", val: "Most deaths in Lethal Company" },
-      ],
-    },
-  ],
-  [
-    MembersEnum.Ipi,
-    {
-      nav: { alt: "Ipi", name: "Iceman Ip", url: "/members/ipi", src: "" },
-      desc: "Ipi is a newcomer in the eyes of RDC, while not yet an official member of RDC yet, one can assume that it's only a matter of time as he has been beating down the crew in Golf and Mario Kart.",
-      stats: [
-        { prop: "", val: "" },
-        { prop: "", val: "" },
-        { prop: "", val: "" },
-      ],
-    },
-  ],
-  [
-    MembersEnum.John,
-    {
-      nav: { alt: "RDC John", name: "J.O.H.N.", url: "/members/john", src: "" },
-      desc: "If you've ever seen or been to DreamCon, this man has had a hand in it. Don't let his sorry gaming skills fool you, John is very valuable in the backend and his contributions should not go unnoticed. Don't forget his immaculate singing ability, John is clearly the best singer in RDC.",
-      stats: [
-        { prop: "", val: "" },
-        { prop: "", val: "" },
-        { prop: "", val: "" },
-      ],
-    },
-  ],
-  [
-    MembersEnum.Aff,
-    {
-      nav: {
-        alt: "RDC Aff",
-        name: "Aff",
-        url: "/members/aff",
-        src: "https://static.wikia.nocookie.net/rdcworld1/images/f/f7/DtlKwRJW4AI3qrN_Aff.jpg/revision/latest?cb=20191004012842",
-      },
-      desc: "Aff and his turtle bring a lot to RDC. Give aff for giving us all the laughs, memes, and behind the scenes work that he contributes to.",
-      stats: [
-        { prop: "", val: "" },
-        { prop: "", val: "" },
-        { prop: "", val: "" },
-      ],
-    },
-  ],
-  [
-    MembersEnum.Lee,
-    {
-      nav: {
-        alt: "RDC Leland",
-        name: "Meland",
-        url: "/members/lee",
-        src: "https://static.wikia.nocookie.net/rdcworld1/images/e/ee/Leland-manigo-image.jpg/revision/latest?cb=20240119040253",
-      },
-      desc: "I love it when you talk to me, my cash machine, my cash machine. The old man laugh always get us. While Leland is really sorry at most games, don't try to play him in any fighting game. You are most likely to get beat up.",
-      stats: [
-        { prop: "", val: "" },
-        { prop: "", val: "" },
-        { prop: "", val: "" },
-      ],
-    },
-  ],
-  [
-    MembersEnum.Dylan,
-    {
-      nav: {
-        alt: "RDC Dylan",
-        name: "The Big Dyl",
-        src: "",
-        url: "/members/dylan",
-      },
-      desc: "The tech guru, the thriller in manilla, el luchador. Dylan is a huge component in the puzzle that is RDC. The flawless streams are thanks to Dylan. He is one of the top gamers in the group although recently he has been falling lower in the ranking of rdc members.",
-      stats: [
-        { prop: "", val: "" },
-        { prop: "", val: "" },
-        { prop: "", val: "" },
-      ],
-    },
-  ],
-  [
-    MembersEnum.Ben,
-    {
-      nav: {
-        alt: "RDC Ben",
-        name: "LaBen James",
-        url: "/members/ben",
-        src: "https://static.wikia.nocookie.net/rdcworld1/images/0/0a/Ben.jpg/revision/latest?cb=20240119050707",
-      },
-      desc: "Ben is a very average gamer. Somethings he is good at and other's hes not. It's hard to explain. Ben gets it done. He works well and if he leaves again he's dead to us.",
-      stats: [
-        { prop: "", val: "" },
-        { prop: "", val: "" },
-        { prop: "", val: "" },
-      ],
-    },
-  ],
-  [
-    MembersEnum.Des,
-    {
-      nav: {
-        alt: "RDC Des",
-        name: "Old Man Des",
-        url: "/members/des",
-        src: "https://static.wikia.nocookie.net/rdcworld1/images/6/62/Desmond-johnson-4.jpg/revision/latest?cb=20191004011638",
-      },
-      desc: "How greeeeeeeat is our God. Ranking as the 2nd best singer in RDC Desmond is surprisingly a very good gamer. You would think the dementia would affect Des but the rage and enthusiasm he plays with always uplifts him in the end. Also Desmond, how is Crystal?",
-      stats: [
-        { prop: "", val: "" },
-        { prop: "", val: "" },
-        { prop: "", val: "" },
-      ],
-    },
-  ],
-]);
 
 /**
  * Returns navigation data for all games.
@@ -234,17 +105,6 @@ export const gameImages = {
   [GamesEnum.CallOfDuty]: "callofduty.jpeg",
 };
 
-export const memberImages = new Map<MembersEnum, string>([
-  [capitalizeFirst(MembersEnum.Mark), "mark_rdc.jpg"],
-  [capitalizeFirst(MembersEnum.Dylan), "dylan_rdc.jpg"],
-  [capitalizeFirst(MembersEnum.Ben), "ben_rdc.jpg"],
-  [capitalizeFirst(MembersEnum.Lee), "leland_rdc.jpg"],
-  [capitalizeFirst(MembersEnum.Des), "desmond_rdc.jpg"],
-  [capitalizeFirst(MembersEnum.John), "john_rdc.jpg"],
-  [capitalizeFirst(MembersEnum.Aff), "aff_rdc.jpg"],
-  [capitalizeFirst(MembersEnum.Ipi), "ipi_rdc.jpg"],
-]);
-
 export const statDescriptions: { [key in $Enums.StatName]: string } = {
   [$Enums.StatName.MK8_DAY]: "Mario Kart 8 Days",
   [$Enums.StatName.MK8_POS]: "Mario Kart 8 Position",
@@ -275,12 +135,6 @@ export enum VisionResultCodes {
   Failed = "Failed",
   CheckRequest = "CheckReq",
 }
-
-type MembersProps = {
-  desc: string;
-  nav: { alt: string; name: string; url: string; src: string };
-  stats: { prop: string; val: string }[]; // TODO: Grab stats from database
-};
 
 // Enhanced game configs with stat references
 export type GameType = "TEAM" | "SOLO";
